Fixing typing bugs

+++ src/Multifractal.jl
@@ -4,15 +4,15 @@ module Multifractal
 
 using Vega
 
-function ChhabraJansen(inputfile::ASCIIString, extensionDq::ASCIIString, extensionFa::ASCIIString, initialQ::Float32, finalQ::Float32, dq::Float32, Np::Int32, r2dq::Float32, r2fa::Float32, scalesToRemove::Int32)
+function ChhabraJansen(inputfile::ASCIIString, extensionDq::ASCIIString, extensionFa::ASCIIString, initialQ::Float64, finalQ::Float64, dq::Float64, Np::Int64, r2dq::Float64, r2fa::Float64, scalesToRemove::Int64)
     println("Chhabra-Jansen multifractal method!");
 end
 
-function fitting(vx::Array{Float32,1}, vy::Array{Float32,1}, N::Int32)
+function fitting(vx::Array{Float64,1}, vy::Array{Float64,1}, N::Int64)
     println("Fitting function");
 end
 
-function calcSumM(x::Array{Float32,1}, y::Array{Float32,1}, Ei::Float32, Ef::Float32, N::Int32)
+function calcSumM(x::Array{Float64,1}, y::Array{Float64,1}, Ei::Float64, Ef::Float64, N::Int64)
     println("CalcSum function");
 end
